DatabaseWrapper.Core.DatabaseHelperBase: New method: AddParameters.
