+++ src/DatabaseWrapper.Core/DatabaseClientBase.cs
@@ -2,12 +2,15 @@
 using System.Collections.Generic;
 using System.Collections.Concurrent;
 using System.Data;
+using static System.FormattableString;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using DatabaseWrapper.Core;
 using ExpressionTree;
 using System.Threading;
+using System.Data.Common;
+using System.Runtime.InteropServices;
 
 namespace DatabaseWrapper.Core
 {
